Add devil role (werewolf detective)
Like the detective but on the side of werewolfs. However, in the code for now, it will be on the side of the villagers. But, for counting who wins, the role does count as a villager, if no werewolves left, the devil loses. In the future, if we count werewolf wins on a user profile or something, we could have an additional function which shows it on the side of the werewolfs.