Reduce remaining use of item if action uses it
At the end of the phase, we reduce items if the corresponding aciton was used by that player.

This happens at the end of the phase, so users can have chance to change their action whilst playing.

Even if the action fails, for examle by a heal target, the item is still used.