Allow actions to be done if player dies that turn
The fact that the witch was not able to resurrect players on the turn the witch dies has frustrated a number of players. Also, the fact that actions done the night a player was killed is also a frustration, the players expected to see the result of their inspect action if they get revived.

This allows all actions to occur without a alive check, We check when the action is done to see if the player is alive, that should be enough.