Fix win check error
When a player was killed by means other than the town lynch or werewolf kill, the game would have the incorrect win check, as the win check was done before the other deaths were calculated. The game would continue one extra phase even if though it was already won. This moves the win check to the end of the end phase function.