Fix issue with overrule action
Overrule action was not being passed as a kill_target when the votes were tied or there were not votes, as the :none target pattern matching took precedence over overrule. The functions were re-ordered to correct the behaviour. The fool win was also added to the overrule version of the kill_player function, as when a fool is burned, the fool wins.